requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
asyncssh==2.6.0
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
black==21.5b1
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
idna==3.1
multidict==5.1.0
mypy-extensions==0.4.3
paramiko==2.7.2
pathspec==0.8.1
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
PyYAML==5.4.1
regex==2021.4.4
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
yarl==1.6.3
